+++ changelog.md
@@ -1,3 +1,7 @@
+## 0.9.0
+CollectionRecord for collectionId and collectionName
+values should be string, not string | undefined since it will just provide an empty string
+
 ## 0.8.0
 use multiple and required in FileField to determine either string string[] or undefined
 take required into account for all fields that have it